+++ dask/array/backends.py
@@ -1,10 +1,13 @@
 import numpy as np
 
+from dask.array.percentile import _percentile
+
 from .dispatch import (
     concatenate_lookup,
     divide_lookup,
     einsum_lookup,
     empty_lookup,
+    percentile_dispatch,
     tensordot_lookup,
 )
 from .numpy_compat import divide as np_divide
@@ -17,6 +20,11 @@ empty_lookup.register((object, np.ndarray), np.empty)
 empty_lookup.register(np.ma.masked_array, np.ma.empty)
 divide_lookup.register((object, np.ndarray), np_divide)
 divide_lookup.register(np.ma.masked_array, ma_divide)
+
+
+@percentile_dispatch.register(np.ndarray)
+def percentile(a, q, interpolation="linear"):
+    return _percentile(a, q, interpolation)
 
 
 @concatenate_lookup.register(np.ma.masked_array)
@@ -108,8 +116,8 @@ def _tensordot(a, b, axes=2):
 def register_cupy():
     import cupy
 
+    from dask.array.dispatch import percentile_dispatch
     from dask.dataframe.backends import percentile
-    from dask.dataframe.dispatch import percentile_dispatch
 
     concatenate_lookup.register(cupy.ndarray, cupy.concatenate)
     tensordot_lookup.register(cupy.ndarray, cupy.tensordot)
