+++ dask/array/dispatch.py
@@ -11,3 +11,8 @@ tensordot_lookup = Dispatch("tensordot")
 einsum_lookup = Dispatch("einsum")
 empty_lookup = Dispatch("empty")
 divide_lookup = Dispatch("divide")
+percentile_dispatch = Dispatch("percentile_dispatch")
+
+
+def _percentile(a, q, interpolation="linear"):
+    return percentile_dispatch(a, q, interpolation)
