+++ dask/array/percentile.py
@@ -91,7 +91,7 @@ def percentile(a, q, interpolation="linear", method="default"):
     --------
     numpy.percentile : Numpy's equivalent Percentile function
     """
-    from dask.dataframe.dispatch import _percentile
+    from dask.array.dispatch import _percentile
 
     from .utils import array_safe, meta_from_array
