+++ dask/dataframe/core.py
@@ -6009,8 +6009,8 @@ def quantile(df, q, method="default"):
         }
     else:
 
+        from dask.array.dispatch import _percentile
         from dask.array.percentile import merge_percentiles
-        from dask.dataframe.dispatch import _percentile
 
         # Add 0 and 100 during calculation for more robust behavior (hopefully)
         calc_qs = np.pad(qs, 1, mode="constant")
