+++ dask/dataframe/dispatch.py
@@ -8,6 +8,7 @@ import pandas as pd
 
 import dask.array as da
 import dask.dataframe as dd
+from dask.array.dispatch import percentile_dispatch  # noqa: F401
 
 from ..utils import Dispatch
 
@@ -22,11 +23,6 @@ concat_dispatch = Dispatch("concat")
 tolist_dispatch = Dispatch("tolist")
 is_categorical_dtype_dispatch = Dispatch("is_categorical_dtype")
 union_categoricals_dispatch = Dispatch("union_categoricals")
-percentile_dispatch = Dispatch("percentile_dispatch")
-
-
-def _percentile(a, q, interpolation="linear"):
-    return percentile_dispatch(a, q, interpolation)
 
 
 def concat(
